lotto/nextjs-flask/requirements.txt
numpy==1.24.3
Flask==3.0.3
SQLAlchemy==2.0.31
mysql-connector-python==8.4.0
Flask_Cors==4.0.1
